refactor: Move icons and rearrange landing page directories
